Add files via upload
Its noted that i took out my mood's just so then it would only take a string as its text instead for my class Tweet. Then for my normal tweets and important tweet i had to edit it so then it wouldnt take a mood just so it works with the clear button and so on.